Introduce Multithreading
Before increasing the scope of the project, I wanted to iron out some
issues. A major one was that the program would crash after a bit because
of a broken pipe. The step logic blocked rendering, and the OS killed
the program. Now, the stepping logic runs on a different thread.